Split Suggestions into UI/logic components for Storybook
Because the original component made use of `useSuggestions`, there was
no easy way to allow users to pass in the mocked terms and products.
With this split, the UI version also simulates the future where this
component may come from FS-UI.
