+++ src/components/search/Suggestions/Suggestions.tsx
@@ -1,15 +1,12 @@
-import { List as UIList } from '@faststore/ui'
+import { gql } from '@vtex/graphql-utils'
+import { useEffect, useState } from 'react'
+import UISuggestions from 'src/components/ui/Search/Suggestions'
+import { request } from 'src/sdk/graphql/request'
 import type {
   SearchSuggestionsQueryQuery,
   SearchSuggestionsQueryQueryVariables,
 } from '@generated/graphql'
-import { gql } from '@vtex/graphql-utils'
-import type { HTMLAttributes } from 'react'
-import { useEffect, useState } from 'react'
-import Button from 'src/components/ui/Button'
-import { request } from 'src/sdk/graphql/request'
-
-import SuggestionProductCard from '../SuggestionProductCard'
+import type { SuggestionsProps } from 'src/components/ui/Search/Suggestions'
 
 const SearchSuggestionsQuery = gql`
   query SearchSuggestionsQuery($term: String!) {
@@ -23,66 +20,6 @@ const SearchSuggestionsQuery = gql`
     }
   }
 `
-
-function formatSearchTerm(
-  indexSubstring: number,
-  searchTerm: string,
-  suggestion: string
-) {
-  if (indexSubstring === 0) {
-    return searchTerm
-      .split('')
-      .map((char, idx) =>
-        idx === 0 && suggestion.indexOf(char.toUpperCase()) === 0
-          ? char.toUpperCase()
-          : char.toLowerCase()
-      )
-      .join('')
-  }
-
-  return searchTerm.toLowerCase()
-}
-
-function handleSuggestions(suggestion: string, searchTerm: string) {
-  const suggestionSubstring = suggestion
-    .toLowerCase()
-    .split(searchTerm.toLowerCase())
-
-  return (
-    <p>
-      {suggestionSubstring.map((substring, indexSubstring) => (
-        <>
-          {substring.length > 0 && (
-            <b data-fs-search-suggestion-item-bold>
-              {indexSubstring === 0
-                ? substring.charAt(0).toUpperCase() + substring.slice(1)
-                : substring}
-            </b>
-          )}
-          {indexSubstring !== suggestionSubstring.length - 1 &&
-            formatSearchTerm(indexSubstring, searchTerm, suggestion)}
-        </>
-      ))}
-    </p>
-  )
-}
-
-export interface SuggestionsProps extends HTMLAttributes<HTMLDivElement> {
-  /**
-   * ID to find this component in testing tools (e.g.: cypress, testing library, and jest).
-   */
-  testId?: string
-  /**
-   * Search term
-   */
-  term?: string
-  /**
-   * Callback to be executed when a suggestion is selected.
-   *
-   * @memberof SuggestionsProps
-   */
-  onSearch: (term: string) => void
-}
 
 function useSuggestions(term: string) {
   const [suggestions, setSuggestions] =
@@ -110,12 +47,7 @@ function useSuggestions(term: string) {
   return { terms, products, loading }
 }
 
-function Suggestions({
-  testId = 'suggestions',
-  term = '',
-  onSearch,
-  ...otherProps
-}: SuggestionsProps) {
+function Suggestions({ term = '', ...otherProps }: SuggestionsProps) {
   const { terms, products, loading } = useSuggestions(term)
 
   if (term.length === 0 && !loading) {
@@ -127,32 +59,12 @@ function Suggestions({
   }
 
   return (
-    <section data-testid={testId} data-fs-search-suggestions {...otherProps}>
-      {terms.length > 0 && (
-        <UIList data-fs-search-suggestion-section>
-          {terms?.map((suggestion) => (
-            <li key={suggestion} data-fs-search-suggestion-item>
-              <Button onClick={() => onSearch(suggestion)}>
-                {handleSuggestions(suggestion, term)}
-              </Button>
-            </li>
-          ))}
-        </UIList>
-      )}
-
-      {products.length > 0 && (
-        <div data-fs-search-suggestion-section>
-          <p data-fs-search-suggestion-title="small">Suggested Products</p>
-          <UIList>
-            {products.map((product, index) => (
-              <li key={product.name} data-fs-search-suggestion-item>
-                <SuggestionProductCard product={product} index={index} />
-              </li>
-            ))}
-          </UIList>
-        </div>
-      )}
-    </section>
+    <UISuggestions
+      term={term}
+      terms={terms}
+      products={products}
+      {...otherProps}
+    />
   )
 }
