+++ src/components/search/Suggestions/index.ts
@@ -2,4 +2,3 @@ export { default as SuggestionsTopSearch } from './SuggestionsTopSearch'
 export type { SuggestionsTopSearchProps } from './SuggestionsTopSearch'
 
 export { default } from './Suggestions'
-export type { SuggestionsProps } from './Suggestions'
