+++ src/components/ui/Search/Suggestions/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Suggestions'
+export type { SuggestionsProps } from './Suggestions'
